feat: Widget テスト用のフィールドを TripApp に追加。ファイル名を変更。
