+++ lib/main.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import 'config/trip_app.dart';
+import 'config/app.dart';
 
 void main() {
   runApp(
-    const ProviderScope(child: TripApp()),
+    const TripApp(),
   );
 }
